Move App to own folder
