+++ src/components/App/App.js
@@ -5,14 +5,14 @@ import {bindActionCreators} from 'redux';
 
 
 //Importing Other Componets
-import Login from './Login/Login';
-import Nav from './Nav/Nav';
-import Favorites from './Favorites/Favorites';
-import Main from './Main/Main'
-import MoviePage from './MoviePage/MoviePage';
-import {getMovies, getUpcomingMovies, getUserFavorites, deleteFavorite, postFavorite} from '../util/apiCalls';
-import {setMovies, setUpcomingMovies,  setFavorites} from '../actions';
-import MovieList from './MovieList/MovieList'
+import Login from '../Login/Login';
+import Nav from '../Nav/Nav';
+import Favorites from '../Favorites/Favorites';
+import Main from '../Main/Main'
+import MoviePage from '../MoviePage/MoviePage';
+import {getMovies, getUpcomingMovies, getUserFavorites, deleteFavorite, postFavorite} from '../../util/apiCalls';
+import {setMovies, setUpcomingMovies,  setFavorites} from '../../actions';
+import MovieList from '../MovieList/MovieList'
 import './App.css';
 import { connect } from 'react-redux';
