+++ src/index.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import rootReducer from './reducers';
-import App from './components/App';
+import App from './components/App/App';
 import { BrowserRouter as Router } from 'react-router-dom'
 import './index.css';
 import * as serviceWorker from './serviceWorker';
